% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space-guide.R
\name{spaceout_style}
\alias{spaceout_style}
\title{The more spaceful style guide}
\usage{
spaceout_style()
}
\description{
The more spaceful style guide
}
\examples{
code <- "say(hi)"
styler::style_text(code, style = spaceout::spaceout_style)
}
